GH-47 -- turned off gestures to swipe back to logout
